API-Security-Tests/requirements.txt
karate
python-owasp-zap-v2.4